pyproject.toml
[tool.poetry]
name = "chat-dataweek-de"
version = "0.1.0"
description = ""
authors = ["Natanael Arndt <natanael.arndt@eccenca.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "chat"}]

[tool.poetry.dependencies]
python = "^3.11"
matrix-nio = {extras = ["e2e"], version = "^0.20.2"}
gunicorn = "^20.1.0"
loguru = "^0.7.0"
python-socketio = "^5.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
